Add integration test for HMS
Spin up a Hive Metastore service in a Github workflow and run integration tests
against it.

The integration test right now is just a little smoke test to prove the
integration tests work. We can add real tests in a subsequent commit.

I ended up having to roll my own Hive Metastore Docker image, which can be found
here:

https://github.com/criccomini/hive-metastore-standalone

The integration test connects to this HMS service and uses it for tests.