More integration tests
Basic tests for fetching tables and columns. At least for the basic
types it seems to work fine. I'll be adding more tests to cover all the
different types but it should be ok to start playing around with
converting the types into Recap. Btw, there's a failing assertion in the
tests right now but check my commit comments and the source code for
that.