Make `test_get_table` work with `EXTERNAL_TABLE`s
A user reported a bug with `EXTERNAL_TABLE`s. HMS was responding with
`tableType = 'MANAGED_TABLE'`. It turns out, we already had a test for this,
which we'd marked as `xfail` for this very reason.

This PR does a number of things:

* Updates hive metastore standalone to use Hive 3.1.3
* Update tests to set `parameters={"EXTERNAL": "TRUE"}`

I did the HMS upgrade because I wanted to verify that the HMS bug hadn't been
fixed later. The tests still failed, so I moved forward with the test fixes.
Users will need to know to set `parameters={"EXTERNAL": "TRUE"}` when creating
external tables. They can see this by looking at the tests.

I suspect the issue is really with a gap in the Thrift protocol. I suspect Hive
Metastore expects `parameters` to have the `EXTERNAL` field set for external
tables. The Python client is, however, just code gen'd from the .thrift files,
which doesn't automatically set `EXTERNAL` when the `tableType` is set. I don't
want to modify the code gen'd code; this is a big no-no. As a result, I'm
leaving it as-is. Users will just have a rouge edge to deal with.

Closes #37